Move shared utils to src/utils.tsx
