+++ src/utils.tsx
@@ -4,11 +4,17 @@ import filesize from 'filesize';
 import purify from 'dompurify';
 import { History, Location } from 'history';
 import queryString from 'query-string';
-import { ChangeInfo } from 'react-diff-view';
+import log from 'loglevel';
+import {
+  Hunks,
+  ChangeInfo,
+  ChangeType,
+  DiffInfo,
+  getChangeKey,
+} from 'react-diff-view';
 
-import { getCodeLineAnchor } from './components/CodeView/utils';
-import { ForwardComparisonMap } from './pages/Compare/utils';
 import { CompareInfo } from './reducers/versions';
+import { getCodeLineAnchor } from './components/CodeView/utils';
 
 // Querystring params used by the app.
 export const messageUidQueryParam = 'messageUid';
@@ -86,18 +92,6 @@ export const createAdjustedQueryString = (
   return `?${queryString.stringify({ ...query, ...newParams })}`;
 };
 
-export const createCodeLineAnchorGetter = ({
-  compareInfo,
-}: {
-  compareInfo: CompareInfo | null | undefined;
-}) => {
-  if (compareInfo && compareInfo.diff) {
-    const map = new ForwardComparisonMap(compareInfo.diff);
-    return map.createCodeLineAnchorGetter();
-  }
-  return getCodeLineAnchor;
-};
-
 export const extractNumber = (text: string): number | null => {
   const matches = text.match(/\d+/);
   if (Array.isArray(matches)) {
@@ -168,4 +162,100 @@ export const codeCanBeHighlighted = ({
   }
 
   return true;
+};
+
+export const getAllHunkChanges = (hunks: Hunks): ChangeInfo[] => {
+  return hunks.reduce(
+    (result: ChangeInfo[], { changes }) => [...result, ...changes],
+    [],
+  );
+};
+
+type ForwardChangeMap = {
+  [line: string]: ChangeInfo[];
+};
+
+const appendChangesInPlace = (
+  changeMap: ForwardChangeMap,
+  line: string,
+  change: ChangeInfo,
+) => {
+  // This destructively appends changes to the line's ChangeInfo array
+  // (when it exists).
+  // eslint-disable-next-line no-param-reassign
+  changeMap[line] = changeMap[line] || [];
+  changeMap[line].push(change);
+};
+
+// This is a map of an old vs. new file comparison, giving preference to the
+// "forward" file, i.e. the newer file.
+export class ForwardComparisonMap {
+  private changeMap: ForwardChangeMap;
+
+  constructor(diff: DiffInfo) {
+    this.changeMap = {};
+    for (const change of getAllHunkChanges(diff.hunks)) {
+      appendChangesInPlace(
+        this.changeMap,
+        String(change.oldLineNumber),
+        change,
+      );
+      appendChangesInPlace(
+        this.changeMap,
+        String(change.newLineNumber),
+        change,
+      );
+    }
+
+    const relevantTypes: ChangeType[] = ['insert', 'normal', 'delete'];
+
+    // Sort changes per line, ordered by relevantTypes.
+    Object.keys(this.changeMap).forEach((line) => {
+      this.changeMap[line].sort((firstChange, secondChange) => {
+        if (firstChange.type === secondChange.type) {
+          return 0;
+        }
+
+        for (const type of relevantTypes) {
+          if (firstChange.type === type) {
+            return -1;
+          }
+          if (secondChange.type === type) {
+            return 1;
+          }
+        }
+
+        // This isn't possible but it needs to return a value to satisfy the
+        // sort() interface
+        /* istanbul ignore next */
+        return 0;
+      });
+    });
+  }
+
+  createCodeLineAnchorGetter() {
+    return this.getCodeLineAnchor.bind(this);
+  }
+
+  getCodeLineAnchor(line: number) {
+    const lineChanges = this.changeMap[line];
+    if (!lineChanges || lineChanges.length === 0) {
+      log.warn(`No changes were mapped for line ${line}`);
+      return '';
+    }
+
+    return `#${getChangeKey(lineChanges[0])}`;
+  }
+}
+
+export const createCodeLineAnchorGetter = ({
+  compareInfo,
+}: {
+  compareInfo: CompareInfo | null | undefined;
+}) => {
+  if (compareInfo && compareInfo.diff) {
+    const map = new ForwardComparisonMap(compareInfo.diff);
+    return map.createCodeLineAnchorGetter();
+  }
+  return getCodeLineAnchor;
 };
